+++ tests/fXmlRpc/Integration/AbstractCallClientIntegrationTest.php
@@ -134,21 +134,4 @@ abstract class AbstractCallClientIntegrationTest extends AbstractIntegrationTest
             $this->assertSame(123, $e->getFaultCode());
         }
     }
-
-    /** @dataProvider getClientsOnly */
-    public function testServerReturnsInvalidResult(Client $client)
-    {
-        $client->setUri(static::$errorEndpoint);
-
-        try {
-            $client->call('system.failure');
-            $this->fail('Exception expected');
-        } catch (\fXmlRpc\Exception\HttpException $e) {
-            $this->assertInstanceOf('fXmlRpc\Exception\TransportException', $e);
-            $this->assertInstanceOf('fXmlRpc\Exception\ExceptionInterface', $e);
-            $this->assertInstanceOf('RuntimeException', $e);
-            $this->assertStringStartsWith('An HTTP error occurred', $e->getMessage());
-            $this->assertSame(500, $e->getCode());
-        }
-    }
 }
