+++ tests/fXmlRpc/Integration/PythonCallClientIntegrationTest.php
@@ -22,6 +22,7 @@
  * SOFTWARE.
  */
 namespace fXmlRpc\Integration;
+use fXmlRpc\Client;
 
 /**
  * @large
@@ -37,4 +38,21 @@ class PythonCallClientIntegrationTest extends AbstractCallClientIntegrationTest
     protected static $command = 'exec python server.py';
 
     protected $disabledExtensions = array('php_curl');
+
+    /** @dataProvider getClientsOnly */
+    public function testServerReturnsInvalidResult(Client $client)
+    {
+        $client->setUri(static::$errorEndpoint);
+
+        try {
+            $client->call('system.failure');
+            $this->fail('Exception expected');
+        } catch (\fXmlRpc\Exception\HttpException $e) {
+            $this->assertInstanceOf('fXmlRpc\Exception\TransportException', $e);
+            $this->assertInstanceOf('fXmlRpc\Exception\ExceptionInterface', $e);
+            $this->assertInstanceOf('RuntimeException', $e);
+            $this->assertStringStartsWith('An HTTP error occurred', $e->getMessage());
+            $this->assertSame(500, $e->getCode());
+        }
+    }
 }
